Luna now using static linking, removed warnings

+++ Luna/src/Luna/ImGui/ImGuiLayer.cpp
@@ -70,7 +70,8 @@ void ImGuiLayer::end()
 {
     ImGuiIO& io = ImGui::GetIO();
     auto& app = Application::get();
-    io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+    io.DisplaySize =
+        ImVec2(static_cast<float>(app.getWindow().getWidth()), static_cast<float>(app.getWindow().getHeight()));
 
     // Rendering
     ImGui::Render();
